requirements.txt
fastmcp==1.0.0
httpx>=0.27.0
uvicorn
numpy
fastapi==0.109.2
# python-dotenv==1.0.1
# PyJWT
# cryptography
# websockets